add -t argument to control which channels are chosen for reformatting
* same as ants_registration.sh:
  * 0 is for no reformatting
  * 1 only reformats the counterstain channel
  * 2 transforms all of the channels

* to consider:
It might be worth only transferring the counterstain channel when 0 or 1 are chosen. This would require changes in this script and probably wouldn't change anything in ants_registration.sh. On the other hand, right now it just slows things down a tiny amount, without impacting the performance of the actual registrations